python/clip_service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
torch>=2.0.0
open_clip_torch>=2.24.0
transformers>=4.41.0
Pillow>=9.4.0
requests>=2.31.0